+++ modules/hanna-css/src/css/CheckboxButton.css.ts
@@ -1,19 +1,21 @@
 import { css } from 'es-in-css';
 
-import { TogglerButtonsGroup, TogglerButtonsKnob } from './styles/forms';
+import { TogglerButtonsKnob } from './styles/forms';
 
+import { CheckboxButtonsGroup_css } from './CheckboxButtonsGroup.css';
 import { RadioButtonsGroup_css } from './RadioButtonsGroup.css';
 
 export default css`
   /*!@deps
     FormField
   */
+
   @media screen {
     ${TogglerButtonsKnob('CheckboxButton')}
-    ${TogglerButtonsGroup('CheckboxButtonsGroup')}
   }
 
   // ===========================================================================
   // Inline for better compression and loading speed
+  ${CheckboxButtonsGroup_css}
   ${RadioButtonsGroup_css}
 `;