+++ modules/hanna-css/src/css/CheckboxButtonsGroup.css.ts
@@ -1,0 +1,17 @@
+import { css } from 'es-in-css';
+
+import { TogglerButtonsGroup } from './styles/forms';
+
+export default css`
+  /*!@deps
+    FormField
+    CheckboxButton
+  */
+`;
+
+// inlined by CheckboxButton.css.ts
+export const CheckboxButtonsGroup_css = () => css`
+  @media screen {
+    ${TogglerButtonsGroup('CheckboxButtonsGroup')}
+  }
+`;