+++ modules/hanna-css/src/css/RadioButtonsGroup.css.ts
@@ -5,11 +5,11 @@ import { TogglerButtonsGroup, TogglerButtonsKnob } from './styles/forms';
 export default css`
   /*!@deps
     FormField
-    CheckboxButtonsGroup
+    CheckboxButton
   */
 `;
 
-// inlined by CheckboxButtonsGroup.css.ts
+// inlined by CheckboxButton.css.ts
 export const RadioButtonsGroup_css = () => css`
   @media screen {
     ${TogglerButtonsKnob('RadioButton')};