+++ modules/hanna-react/src/CheckboxButton.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { TogglerInput, TogglerInputProps } from './_abstract/_TogglerInput';
+
+export type CheckboxProps = TogglerInputProps;
+
+export const CheckboxButton = (props: CheckboxProps) => (
+  <TogglerInput bem="CheckboxButton" {...props} type="checkbox" innerWrap />
+);
+
+export default CheckboxButton;